! update after release `v1.0.0-Alpha4`

+++ scripts/maven_deploy.sh
@@ -9,9 +9,11 @@ jvu::switch_to_jdk 21
 ########################################
 # integration test
 ########################################
-rm -rf "$HOME/.m2/repository/io/foldright"/cffu*
-cu::log_then_run scripts/integration_test
-cu::log_then_run demos/scripts/integration_test
+if [ "${1:-}" = -it ]; then
+  rm -rf "$HOME/.m2/repository/io/foldright"/cffu*
+  cu::log_then_run scripts/integration_test
+  cu::log_then_run demos/scripts/integration_test
+fi
 
 ########################################
 # maven deploy
